+++ src/main/java/com/github/kd_gaming1/packcore/gui/screen/WelcomeWizardScreen.java
@@ -71,6 +71,9 @@ public class WelcomeWizardScreen extends AbstractScreen {
         navigator.addPage(new TabDesignPage(wizardState, navigator, contentWidth, contentHeight));
         navigator.addPage(new ItemBackgroundPage(wizardState, navigator, contentWidth, contentHeight));
         navigator.addPage(new StorageDesignPage(wizardState, navigator, contentWidth, contentHeight));
+        if (FabricLoader.getInstance().isModLoaded("secretroutesmod")) {
+            navigator.addPage(new DungeonRoutesPage(wizardState, navigator, contentWidth, contentHeight));
+        }
         if (FabricLoader.getInstance().isModLoaded("scaleme")) {
             navigator.addPage(new SwordBlockPage(wizardState, navigator, contentWidth, contentHeight));
         }