+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 ### Changes
 - Update the performance profile to edit More Culling options too for better fps.
+- Added Secret Routes page to welcome wizard.